requirements.txt
chromadb
sentence-transformers
rdflib
faiss-cpu
numpy
